added data consistency checks and no entry case

+++ companion/tracking.js
@@ -1,6 +1,3 @@
-/**
- * The module is responsible for time tracking
- */
 import {MESSAGE_TYPE} from '../common/message-types';
 import {gettext} from 'i18n';
 import {debug} from '../common/debug';
@@ -11,7 +8,18 @@ const CURRENT_ENTRY_REFRESH_INTERVAL_MS = 15000;
 
 const OPTIMAL_TEXTS_LENGTH = 64;
 
-const entryUniquenessId = ({id, start, stop, pid, description}) => [id, start, stop, pid, description].join(':');
+const NO_PROJECT_INFO = {
+  color: NO_PROJECT_COLOR,
+  projectName: gettext('no_project'),
+};
+
+const EMPTY_TIME_ENTRY = {
+  desc: gettext('no_description'),
+  billable: false,
+  ...NO_PROJECT_INFO,
+};
+
+const entryUniquenessId = ({id, start, stop, pid, description} = {}) => [id, start, stop, pid, description].join(':');
 
 const optimiseStringForDisplaying = (input) => {
   if (input && input.length > OPTIMAL_TEXTS_LENGTH) {
@@ -21,7 +29,7 @@ const optimiseStringForDisplaying = (input) => {
 };
 
 /**
- *
+ * The module is responsible for time tracking process
  */
 class Tracking {
   /**
@@ -46,7 +54,7 @@ class Tracking {
    */
   async initialize() {
     await this._updateUserData();
-    await this._setupCurrentEntryTrack();
+    await this._setupCurrentEntryTracking();
     debug('tracking initialized');
   }
 
@@ -74,55 +82,45 @@ class Tracking {
    * @private
    */
   _attachCommandsProcessing() {
-    this._transmitter.onMessage(MESSAGE_TYPE.STOP_CURRENT_ENTRY, ({stop}) => {
-      this.stopCurrentEntry(stop);
+    this._transmitter.onMessage(MESSAGE_TYPE.STOP_CURRENT_ENTRY, async ({stop, id}) => {
+      this.currentEntry = {
+        ...this.currentEntry,
+        stop: new Date(stop).toISOString(),
+      };
+
+      if (id === this.currentEntryId) {
+        await this._api.updateTimeEntry(this.currentEntry);
+      }
+
+      await this._setupCurrentEntryTracking();
     });
 
-    this._transmitter.onMessage(MESSAGE_TYPE.RESUME_LAST_ENTRY, ({start}) => {
-      this.resumeLastEntry(start);
+    this._transmitter.onMessage(MESSAGE_TYPE.RESUME_LAST_ENTRY, async ({start, id}) => {
+      this.currentEntry = {
+        wid: this.user && this.user.default_workspace_id,
+        ...this.currentEntry,
+        stop: null,
+        start: new Date(start).toISOString(),
+      };
+
+      if (id === this.currentEntryId) {
+        await this._api.createTimeEntry(this.currentEntry);
+      }
+
+      await this._setupCurrentEntryTracking();
     });
 
-    this._transmitter.onMessage(MESSAGE_TYPE.DELETE_CURRENT_ENTRY, () => {
-      this.deleteCurrentEntry();
+    this._transmitter.onMessage(MESSAGE_TYPE.DELETE_CURRENT_ENTRY, async ({id}) => {
+      if (id === this.currentEntryId) {
+        await this._api.deleteTimeEntry(this.currentEntry);
+      }
+      this.currentEntry = null;
+      await this._setupCurrentEntryTracking();
     });
   }
 
-  /**
-   * Deletes current entry
-   */
-  async deleteCurrentEntry() {
-    await this._api.deleteTimeEntry(this.currentEntry);
-    this.currentEntry = null;
-    await this._setupCurrentEntryTrack();
-  }
-
-  /**
-   * Stops current entry
-   * @param {number} stop
-   */
-  async stopCurrentEntry(stop) {
-    this.currentEntry = {
-      ...this.currentEntry,
-      stop: new Date(stop).toISOString(),
-    };
-
-    await this._api.updateTimeEntry(this.currentEntry);
-    await this._setupCurrentEntryTrack();
-  }
-
-  /**
-   * Starts a new entry based on current/last
-   * @param {number} start
-   */
-  async resumeLastEntry(start) {
-    this.currentEntry = {
-      ...this.currentEntry,
-      stop: null,
-      start: new Date(start).toISOString(),
-    };
-
-    await this._api.createTimeEntry(this.currentEntry);
-    await this._setupCurrentEntryTrack();
+  get currentEntryId() {
+    return this.currentEntry && this.currentEntry.id;
   }
 
   /**
@@ -130,7 +128,7 @@ class Tracking {
    * @return {Promise<void>}
    * @private
    */
-  async _setupCurrentEntryTrack() {
+  async _setupCurrentEntryTracking() {
     if (this._updateCurrentEntryInterval) {
       clearInterval(this._updateCurrentEntryInterval);
     }
@@ -173,35 +171,55 @@ class Tracking {
 
   /**
    * Getter for current entry info to be transferred to the device
-   * @return {{id: string}|null}
+   * @return {{id: *, billable: boolean, desc: string}}
    * @private
    */
   get _currentEntryMessage() {
     if (!this.currentEntry) {
-      return null;
+      return EMPTY_TIME_ENTRY;
     }
 
-    const project = this.projects.find(({id}) => id === this.currentEntry.pid);
+    const start = this._currentEntryStartTimestamp;
 
-    const noProject = {
-      color: NO_PROJECT_COLOR,
-      projectName: gettext('no_project'),
-    };
-
-    const projectInfo = project && {
-      color: project.color,
-      projectName: optimiseStringForDisplaying(project.name),
-    } || noProject;
-
-    const start = !this.currentEntry.stop && Date.parse(this.currentEntry.start);
+    const desc = optimiseStringForDisplaying(this.currentEntry.description) || gettext('no_description');
 
     return {
-      id: this.currentEntry.id,
-      desc: optimiseStringForDisplaying(this.currentEntry.description),
+      id: this.currentEntryId,
       billable: this.currentEntry.billable,
+      desc,
       ...(start ? {start} : {}),
-      ...projectInfo,
+      ...this._currentEntryMessageProjectInfo,
     };
+  }
+
+  /**
+   * Returns time in milliseconds from entry start in case if entry still running
+   * @return {boolean|number}
+   * @private
+   */
+  get _currentEntryStartTimestamp() {
+    return !this.currentEntry.stop && Date.parse(this.currentEntry.start);
+  }
+
+  /**
+   * Returns information about the project needs to be transported to the device
+   * @return {*|{color: *, projectName: string}|{color: string, projectName: *}}
+   * @private
+   */
+  get _currentEntryMessageProjectInfo() {
+    return this._currentEntryProject && {
+      color: this._currentEntryProject.color,
+      projectName: optimiseStringForDisplaying(this._currentEntryProject.name),
+    } || NO_PROJECT_INFO;
+  }
+
+  /**
+   * returns project having id equals to pid of current entry
+   * @return {*}
+   * @private
+   */
+  get _currentEntryProject() {
+    return this.projects.find(({id}) => id === this.currentEntry.pid);
   }
 }
 
@@ -209,5 +227,6 @@ export {
   Tracking,
   CURRENT_ENTRY_REFRESH_INTERVAL_MS,
   NO_PROJECT_COLOR,
+  NO_PROJECT_INFO,
   OPTIMAL_TEXTS_LENGTH,
 };
